strawberry_fastapi/requirements.txt
fastapi
pydantic
strawberry-graphql[fastapi]
couchdb
uvicorn
# python-jose[cryptography]
# python-multipart
ldap3
requests
Jinja2
pytz